create reciever file and remove from boltServices
